Feature/data generator tests (#3)
* CI updated

- build renamed as build-dev
- redundant dep submission removed

* IDEA plugin configured for all

* kafka consumer starter module created

* tests added for util class

* ci update

* sonar and jacoco config

* sonar and jacoco config

* ci config enhanced
