--- .github/workflows/build.yaml
@@ -1,64 +1,0 @@
-name: Build
-
-on:
-  push:
-    branches:
-      - dev
-jobs:
-  build-jar:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-    outputs:
-      BUILD_NUMBER: ${{ steps.build-number.outputs.BUILD_NUMBER }}
-    steps:
-      - uses: actions/checkout@v4
-      - id: build-number
-        name: Configuring Build Number
-        run: |
-          echo "BUILD_NUMBER=$(git rev-parse --short=8 HEAD)" >> $GITHUB_OUTPUT
-      - name: Test Build Number
-        run: echo "${{ steps.build-number.outputs.BUILD_NUMBER }}"
-      - name: Set up JDK 21
-        uses: actions/setup-java@v4
-        with:
-          java-version: '21'
-          distribution: 'temurin'
-      - name: Setup Gradle
-        uses: gradle/actions/setup-gradle@v3.1.0
-      - name: Build
-        run: ./gradlew clean assemble -Pversion=${{ steps.build-number.outputs.BUILD_NUMBER }}
-      - name: Test
-        run: ./gradlew test
-      - name: Integration Test
-        run: ./gradlew integrationTest
-      - name: Uploading DataGenerator Jars
-        uses: actions/upload-artifact@master
-        with:
-          name: data-generator-jar
-          path: module/app/data-generator/build/libs
-
-  build-data-generator-image:
-    needs:
-      - build-jar
-    name: Data Generator - Build & Push Docker Image
-    runs-on: ubuntu-latest
-    permissions:
-      packages: write
-      contents: read
-    steps:
-      - name: Check out the repo
-        uses: actions/checkout@v4
-
-      - name: Log in to the Container registry
-        uses: docker/login-action@65b78e6e13532edd9afa3aa52ac7964289d1a9c1
-        with:
-          registry: ghcr.io
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-      - name: Build and push Docker images
-        uses: docker/build-push-action@3b5e8027fcad23fda98b2e3ac259d8d67585f671
-        with:
-          context: module/app/data-generator
-          push: true
-          tags: ghcr.io/maeworkshop/data-generator:${{ needs.build-jar.outputs.BUILD_NUMBER }}
