+++ .github/workflows/check.yaml
@@ -1,11 +1,11 @@
-name: Feature/Bugfix Check
+name: Check Branch
 
 on:
   push:
     branches:
       - feature/**
       - bugfix/**
-
+  workflow_dispatch:
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -13,30 +13,29 @@ jobs:
       contents: read
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
       - name: Set up JDK 21
         uses: actions/setup-java@v4
         with:
           java-version: '21'
           distribution: 'temurin'
+      - name: Get current branch name
+        id: get_branch
+        run: echo "CURRENT_BRANCH_NAME=$(git rev-parse --abbrev-ref HEAD)" >> $GITHUB_ENV
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v3.1.0
       - name: Build
-        run: ./gradlew clean assemble
+        run: CI=true ./gradlew clean assemble
       - name: Test
-        run: ./gradlew test
+        run: CI=true ./gradlew test
       - name: Integration Test
-        run: ./gradlew integrationTest
-
-  dependency-submission:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up JDK 21
-        uses: actions/setup-java@v4
-        with:
-          java-version: '21'
-          distribution: 'temurin'
-      - name: Generate and submit dependency graph
-        uses: gradle/actions/dependency-submission@v3.1.0
+        run: CI=true ./gradlew integrationTest
+      - name: Sonar Scan
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}  # Needed to get PR information, if any
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+          BRANCH_NAME: ${{ env.CURRENT_BRANCH_NAME }}
+        run: |
+          echo "Branch name $BRANCH_NAME"
+          CI=true ./gradlew sonar --info -Dsonar.branch.name=$BRANCH_NAME
