+++ build.gradle.kts
@@ -1,5 +1,18 @@
 import org.gradle.plugins.ide.idea.model.IdeaModel
 
+plugins {
+    id("org.sonarqube") version "4.4.1.3373"
+}
+
+sonar {
+    properties {
+        property("sonar.projectKey", "maeworkshop_k8s-workshop")
+        property("sonar.organization", "maeworkshop")
+        property("sonar.host.url", "https://sonarcloud.io")
+        property("sonar.coverage.jacoco.xmlReportPaths", "**/build/reports/jacoco/test/jacocoTestReport.xml")
+    }
+}
+
 subprojects {
     pluginManager.withPlugin("java") {
         val integrationTestSrcPath = "src/integrationTest"
@@ -50,6 +63,39 @@ subprojects {
 
                 named("check") {
                     dependsOn(integrationTestTaskName)
+                }
+            }
+
+            pluginManager.withPlugin("jacoco") {
+                tasks {
+                    named("test") {
+                        finalizedBy(named("jacocoTestReport"))
+                    }
+
+                    named("jacocoTestReport") {
+                        dependsOn("test")
+                    }
+                }
+
+
+            }
+
+            pluginManager.withPlugin("jacoco") {
+                tasks {
+                    named("test") {
+                        finalizedBy(named("jacocoTestReport"))
+                    }
+
+
+                    named("jacocoTestReport") {
+                        dependsOn("test")
+                    }
+
+                    withType<JacocoReport> {
+                        reports {
+                            xml.required = true
+                        }
+                    }
                 }
             }
