--- module/app/data-generator/src/main/java/com/maemresen/data/generator/config/SchedulerConfig.java
@@ -1,10 +1,0 @@
-package com.maemresen.data.generator.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@EnableScheduling
-@Configuration
-public class SchedulerConfig {
-
-}
