--- module/app/data-generator/src/main/java/com/maemresen/data/generator/config/props/SensorDataGeneratorProps.java
@@ -1,32 +1,0 @@
-package com.maemresen.data.generator.config.props;
-
-import java.util.List;
-import com.maemresen.lib.message.dto.SensorType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ConfigurationProperties(prefix = "app.generator.sensor-data")
-@Configuration
-public class SensorDataGeneratorProps {
-
-  private List<LocationProps> locations;
-
-  public record LocationProps(String name, List<DeviceProps> devices) {
-  }
-
-  public record DeviceProps(String name, List<SensorProps> sensors) {
-
-  }
-
-  public record SensorProps(String name, SensorType type) {
-
-  }
-}
