--- module/app/data-generator/src/main/java/com/maemresen/data/generator/util/Topic.java
@@ -1,5 +1,0 @@
-package com.maemresen.data.generator.util;
-
-public enum Topic {
-  SENSOR_DATA
-}
