+++ module/app/k8s-workshop-data-generator/Dockerfile
