+++ module/app/k8s-workshop-data-generator/src/main/java/com/maemresen/k8s/workshop/data/generator/DataGeneratorApp.java
@@ -1,0 +1,17 @@
+package com.maemresen.k8s.workshop.data.generator;
+
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan("com.maemresen")
+@SpringBootApplication
+public class DataGeneratorApp {
+
+  public static void main(String[] args) {
+    new SpringApplicationBuilder(DataGeneratorApp.class)
+        .web(WebApplicationType.NONE)
+        .run(args);
+  }
+}
