+++ module/app/k8s-workshop-data-generator/src/main/resources/application.yaml
