+++ module/app/k8s-workshop-listener/src/main/java/com/maemresen/k8s/workshop/listener/util/Topic.java
@@ -1,0 +1,5 @@
+package com.maemresen.k8s.workshop.listener.util;
+
+public enum Topic {
+  SENSOR_DATA
+}
