+++ module/app/k8s-workshop-listener/src/main/resources/application.yaml
