--- module/app/listener/src/main/java/com/maemresen/listener/util/Topic.java
@@ -1,5 +1,0 @@
-package com.maemresen.listener.util;
-
-public enum Topic {
-  SENSOR_DATA
-}
