+++ module/lib/k8s-workshop-avro-schemes/src/main/avro/sensor_data.avsc
