+++ module/lib/k8s-workshop-messaging-spring-boot-starter/src/main/java/com/maemresen/k8s/workshop/messaging/starter/MessagingAutoConfiguration.java
@@ -1,0 +1,17 @@
+package com.maemresen.k8s.workshop.messaging.starter;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MessagingAutoConfiguration {
+
+  @ConfigurationProperties(prefix = "app.messaging")
+  @ConditionalOnMissingBean
+  @Bean
+  public MessagingProps messagingProps() {
+    return new MessagingProps();
+  }
+}
