+++ module/lib/k8s-workshop-messaging-spring-boot-starter/src/main/java/com/maemresen/k8s/workshop/messaging/starter/Topic.java
@@ -1,0 +1,5 @@
+package com.maemresen.k8s.workshop.messaging.starter;
+
+public enum Topic {
+  SENSOR_DATA
+}
