+++ module/lib/k8s-workshop-messaging-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+com.maemresen.k8s.workshop.messaging.starter.MessagingAutoConfiguration
